requirements.txt
python-dotenv==1.0.0
ollama==0.1.6
streamlit==1.32.0
pandas==2.2.1 